requirements.txt
fastapi==0.78.0
uvicorn==0.15.0
pandas==1.4.2
python-dotenv==0.15.0
schedule==1.1.0